selecty na laborki

+++ struktura.sql
@@ -113,7 +113,7 @@ CREATE TABLE sprzedaz (
   produkt_id INTEGER NOT NULL,
   lokalizacja_id INTEGER NOT NULL,
   ilosc INTEGER NOT NULL,
-  wartosc DECIMAL(12,2) NOT NULL,
+  wartosc INTEGER NOT NULL,
   CONSTRAINT sprzedaz_pk PRIMARY KEY (id),
   CONSTRAINT sprzedaz_klient_fk FOREIGN KEY (klient_id) REFERENCES klient (id),
   CONSTRAINT sprzedaz_sprzedawca_fk FOREIGN KEY (sprzedawca_id) REFERENCES sprzedawca (id),
